Add LatestTransactionListView widget for home screen
This commit introduces the `LatestTransactionListView` widget, designed to display a list of recent transactions on the home screen. The widget is implemented in the `lib/features/presentation/home/view/widgets/latest_transaction_list_view.dart` file. It utilizes a `SingleChildScrollView` with a horizontal scroll direction to display a list of `UserInfoModel` objects, each rendered using the `UsreInfoListTile` widget.

The `UserInfoModel` objects are hardcoded for demonstration purposes, including sample data for three users. This widget is part of the home screen's presentation layer and is intended to enhance the user interface by providing a clear view of recent activities.