Add Transaction Item List View Widget
Create a new widget, `TrasnactionItemListView`, in the home view section to display a list of transaction items. This widget includes a hardcoded list of `TransactionModel` items for demonstration purposes. Each transaction item is wrapped in a `Padding` widget for spacing and uses the `TrasnactionItem` widget to render individual transaction details.